Move Mail Cleaner API to Enterprise Server
