+++ SolidCP/Sources/SolidCP.EnterpriseServer.Code/Servers/ServerController.cs
@@ -2394,6 +2394,10 @@ namespace SolidCP.EnterpriseServer
                 }
             }
 
+            //Add Domain to Mail Cleaner
+            if (domainType == DomainType.Domain && pointMailDomainId != 0) //Only For domain -- Ignore Subdomains 
+                APIMailCleanerHelper.DomainAdd(domainName, packageId);
+
             return domainId;
 		}
 
@@ -2794,7 +2798,10 @@ namespace SolidCP.EnterpriseServer
 				// delete domain
 				DataProvider.DeleteDomain(SecurityContext.User.UserId, domainId);
 
-				return 0;
+                //Delete Domain from Mail Cleaner
+                APIMailCleanerHelper.DomainRemove(domain.DomainName, domain.PackageId);
+
+                return 0;
 			}
 			catch (Exception ex)
 			{
