+++ SolidCP/Sources/SolidCP.WebPortal/DesktopModules/SolidCP/App_Code/IPAddressExtensions.cs
@@ -147,165 +147,4 @@ namespace Knom.Helpers.Net
             return network1.Equals(network2);
         }
     }
-
-    public static class APIMailCleanerHelper  
-    {
-        private static StringDictionary ConvertArrayToDictionary(string[] settings)
-        {
-            StringDictionary r = new StringDictionary();
-            foreach (string setting in settings)
-            {
-                int idx = setting.IndexOf('=');
-                r.Add(setting.Substring(0, idx), setting.Substring(idx + 1));
-            }
-            return r;
-        }
-        public static String getServiceURL(int PlanID)
-        {
-            var l_URL = SolidCP.Portal.ES.Services.Servers.GetMailFilterUrlByHostingPlan(PlanID, ResourceGroups.Filters);
-            if (!String.IsNullOrEmpty(l_URL))
-            {
-                l_URL = l_URL.TrimEnd('/');
-                l_URL += "/api/";
-            }
-            return l_URL;
-        }
-        public static String getServiceURL()
-        {
-            //String l_URL = string.Empty;
-            //var fileter =  SolidCP.Portal.ES.Services.Servers.GetPackageServiceProvider(PanelSecurity.PackageId, ResourceGroups.Filters);
-            var l_URL = SolidCP.Portal.ES.Services.Servers.GetMailFilterUrl(PanelSecurity.PackageId, ResourceGroups.Filters);
-           
-
-
-            //var Session = System.Web.HttpContext.Current.Session;
-            //var l_PackageID  =  Session["currentPackage"];//server id
-            //System.Data.DataSet dsServers = SolidCP.Portal.ES.Services.Servers.GetRawServers();
-            //if (dsServers == null)
-            //    return String.Empty;
-
-            //var l_Services = new System.Data.DataView(dsServers.Tables[1], "ServerID=" + l_PackageID.ToString() , "", System.Data.DataViewRowState.CurrentRows);
-            //if(l_Services.Count == 0)
-            //    return String.Empty;
-
-            ////+ " And ServiceName = 'Mail Cleaner'"
-            //int l_ServiceID = 0;
-            //SolidCP.EnterpriseServer.ServiceInfo l_oServiceInfo = null;
-            //Boolean MailFilterAdded = false;
-            //foreach (System.Data.DataRowView l_oServiceRow in l_Services)
-            //{
-            //    l_ServiceID = Convert.ToInt16( l_oServiceRow["ServiceID"]);
-            //    l_oServiceInfo = SolidCP.Portal.ES.Services.Servers.GetServiceInfo(Convert.ToInt16(l_ServiceID));
-            //    var l_oProvider = SolidCP.Portal.ES.Services.Servers.GetProvider(l_oServiceInfo.ProviderId);
-            //    if (l_oProvider.ProviderName.ToUpper().Equals("MAILCLEANER"))
-            //    {
-            //        MailFilterAdded = true;
-            //        break;
-            //    }
-            //}
-
-            //if (!MailFilterAdded)
-            //    return String.Empty;
-
-
-            //// load service properties and bind them
-            //string[] settings = SolidCP.Portal.ES.Services.Servers.GetServiceSettings(Convert.ToInt16(l_ServiceID));
-            //if (settings == null)
-            //    return String.Empty;
-
-            //var l_ServiceSettings = ConvertArrayToDictionary(settings);
-            //// load resource group details
-            ////var resourceGroup = SolidCP.Portal.ES.Services.Servers.GetResourceGroup(provider.GroupId);
-            //l_URL = l_ServiceSettings["apiurl"];
-
-            if (!String.IsNullOrEmpty(l_URL))
-            { 
-                l_URL = l_URL.TrimEnd('/');
-                l_URL += "/api/";
-            }
-            return l_URL;
-        }
-        public static void APICall(String f_stParam, int f_iPlanID = 0)
-        {
-            String l_URL = string.Empty;
-            try
-            {
-                if(f_iPlanID==0)
-                    l_URL = getServiceURL();
-                else
-                    l_URL = getServiceURL(f_iPlanID);
-            }
-            catch (Exception ex)
-            {
-                throw (ex);
-                 //throw( new Exception("MAILCLEANER_API_404"));
-            } 
-
-            if (String.IsNullOrEmpty(l_URL))
-                return;
-
-            // Set to use TLS1.2
-            ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
-            //"https://10.2.150.107/api/"
-            // Create a request for the URL. 
-            WebRequest request = WebRequest.Create(l_URL + f_stParam);
-            // If required by the server, set the credentials.
-            request.Credentials = CredentialCache.DefaultCredentials;
-            // Get the response.
-            WebResponse response = request.GetResponse();
-            // Display the status.
-            //Console.WriteLine(((HttpWebResponse)response).StatusDescription);
-            // Get the stream containing content returned by the server.
-            Stream dataStream = response.GetResponseStream();
-            // Open the stream using a StreamReader for easy access.
-            StreamReader reader = new StreamReader(dataStream);
-
-            // Display the content.
-            //Console.WriteLine(responseFromServer);
-            if (response != null)
-            {
-                XmlDocument xmlResult = new XmlDocument();
-                // Read the content.
-                String lstXMLResult = reader.ReadToEnd();
-                xmlResult.LoadXml(lstXMLResult);
-                String lstMessage = String.Empty;
-                foreach (XmlNode lxmNode in xmlResult.ChildNodes)
-                {
-                    if (lxmNode.Name == "response")
-                    {
-                        if (lxmNode.FirstChild.Name == "message")
-                            lstMessage = lxmNode.FirstChild.InnerText;
-                    }
-                }
-                // MessageBox.Show(lstMessage);
-            }
-            // Clean up the streams and the response.
-            reader.Close();
-            response.Close();
-
-
-        }
-        public static void DomainAdd(String f_stDomain, int f_iPlanID)
-        {
-            HostingPlanContext cntx = PackagesHelper.GetCachedHostingPlanContext(f_iPlanID);
-            if (cntx != null)
-            { 
-                if (Utils.CheckQouta("Filters.Enable", cntx))
-                {
-                    APICall("domain/add/name/" + f_stDomain, f_iPlanID);
-
-                } 
-            }
-        }
-        public static void DomainAdd(String f_stDomain)
-        {
-            if(PackagesHelper.IsQuotaEnabled(PanelSecurity.PackageId, "Filters.Enable"))
-              APICall("domain/add/name/" + f_stDomain);
-        }
-        public static void DomainRemove(String f_stDomain)
-        {
-            if (PackagesHelper.IsQuotaEnabled(PanelSecurity.PackageId, "Filters.Enable"))
-                APICall("domain/remove/name/" + f_stDomain);
-        }
-    }
 }
