+++ SolidCP/Sources/SolidCP.WebPortal/DesktopModules/SolidCP/DomainsEditDomain.ascx.cs
@@ -232,8 +232,6 @@ namespace SolidCP.Portal
                     ShowResultMessage(result);
                     return;
                 }
-                 //Delete Domain to Mail Cleaner
-                 Knom.Helpers.Net.APIMailCleanerHelper.DomainRemove(DomainName.Text);
             }
             catch (Exception ex)
             {
